+++ .gitignore
@@ -1,13 +1,5 @@
-/**/.vs/
-/**/.vscode/*
-!/**/.vscode/.vscode/cmake-kits.json
-/**/build/
-/**/out/
-/arm-gnu-toolchain/
-/lib/*/crt0.o
-/lib/*/gba-irq.o
-/lib/*/gba-syscalls.o
-/lib/*/gba.specs
-/lib/gbaplusplus*
-/bin/
-/sample_project/data/
+.idea/
+arm-gnu-toolchain/
+lib/tonc/
+tools/
+dependencies.txt
