+++ lib/multiboot/.gitignore
@@ -1,0 +1,4 @@
+cmake-build-debug/
+cmake-build-debug-clang/
+cmake-build-release/
+cmake-build-release-clang/
