+++ lib/rom/gba-syscalls.c
@@ -93,5 +93,3 @@ int _wait(int * status) {
 int _write(int file, char * ptr, int len) {
   return 0;
 }
-
-void __sync_synchronize() {}
