--- sample_project/.vscode/cmake-kits.json
@@ -1,6 +1,0 @@
-[
-    {
-        "name" : "ARM GBA",
-        "toolchainFile" : "arm-gba-toolchain.cmake"
-    }
-]
