--- sample_project/CMakeSettings.json
@@ -1,114 +1,0 @@
-﻿{
-  "configurations": [
-    {
-      "name": "Debug",
-      "generator": "Ninja",
-      "configurationType": "Debug",
-      "buildRoot": "${projectDir}\\out\\build\\${name}",
-      "installRoot": "${projectDir}\\out\\install\\${name}",
-      "cmakeCommandArgs": "",
-      "buildCommandArgs": "-v",
-      "ctestCommandArgs": "",
-      "inheritEnvironments": [],
-      "variables": [
-        {
-          "name": "CMAKE_C_FLAGS_DEBUG",
-          "value": "-O0 -g3",
-          "type": "STRING"
-        },
-        {
-          "name": "CMAKE_C_FLAGS_RELWITHDEBINFO",
-          "value": "-O2 -g -DNDEBUG",
-          "type": "STRING"
-        },
-        {
-          "name": "CMAKE_C_FLAGS_RELEASE",
-          "value": "-O3 -DNDEBUG",
-          "type": "STRING"
-        },
-        {
-          "name": "CMAKE_C_FLAGS_MINSIZEREL",
-          "value": "-Os -DNDEBUG",
-          "type": "STRING"
-        },
-        {
-          "name": "CMAKE_CXX_FLAGS_DEBUG",
-          "value": "-O0  -g3",
-          "type": "STRING"
-        },
-        {
-          "name": "CMAKE_CXX_FLAGS_MINSIZEREL",
-          "value": "-Os -DNDEBUG",
-          "type": "STRING"
-        },
-        {
-          "name": "CMAKE_CXX_FLAGS_RELEASE",
-          "value": "-O3 -DNDEBUG",
-          "type": "STRING"
-        },
-        {
-          "name": "CMAKE_CXX_FLAGS_RELWITHDEBINFO",
-          "value": "-O2 -g -DNDEBUG",
-          "type": "STRING"
-        }
-      ],
-      "intelliSenseMode": "linux-gcc-arm",
-      "cmakeToolchain": "${projectDir}\\..\\arm-gba-toolchain.cmake"
-    },
-    {
-      "name": "Release",
-      "generator": "Ninja",
-      "configurationType": "Release",
-      "buildRoot": "${projectDir}\\out\\build\\${name}",
-      "installRoot": "${projectDir}\\out\\install\\${name}",
-      "cmakeCommandArgs": "",
-      "buildCommandArgs": "-v",
-      "ctestCommandArgs": "",
-      "inheritEnvironments": [],
-      "variables": [
-        {
-          "name": "CMAKE_C_FLAGS_DEBUG",
-          "value": "-g",
-          "type": "STRING"
-        },
-        {
-          "name": "CMAKE_C_FLAGS_RELWITHDEBINFO",
-          "value": "-O2 -g -DNDEBUG",
-          "type": "STRING"
-        },
-        {
-          "name": "CMAKE_C_FLAGS_RELEASE",
-          "value": "-O3 -DNDEBUG",
-          "type": "STRING"
-        },
-        {
-          "name": "CMAKE_C_FLAGS_MINSIZEREL",
-          "value": "-Os -DNDEBUG",
-          "type": "STRING"
-        },
-        {
-          "name": "CMAKE_CXX_FLAGS_DEBUG",
-          "value": "-g",
-          "type": "STRING"
-        },
-        {
-          "name": "CMAKE_CXX_FLAGS_MINSIZEREL",
-          "value": "-Os -DNDEBUG",
-          "type": "STRING"
-        },
-        {
-          "name": "CMAKE_CXX_FLAGS_RELEASE",
-          "value": "-O3 -DNDEBUG",
-          "type": "STRING"
-        },
-        {
-          "name": "CMAKE_CXX_FLAGS_RELWITHDEBINFO",
-          "value": "-O2 -g -DNDEBUG",
-          "type": "STRING"
-        }
-      ],
-      "intelliSenseMode": "linux-gcc-arm",
-      "cmakeToolchain": "${projectDir}\\..\\arm-gba-toolchain.cmake"
-    }
-  ]
-}
