--- sample_project/ProjectSettings.json
@@ -1,3 +1,0 @@
-{
-  "CurrentProjectSetting": "ROM-Debug"
-}
