--- sample_project/launch.vs.json
@@ -1,49 +1,0 @@
-{
-  "version": "0.2.1",
-  "defaults": {},
-  "configurations": [
-    {
-      "type": "cppdbg",
-      "project": "CMakeLists.txt",
-      "projectTarget": "rom.elf",
-      "name": "mGBA Debug",
-      "MIMode": "gdb",
-      "program": "${cmake.buildRoot}/rom.elf",
-      "targetArchitecture": "arm",
-      "miDebuggerServerAddress": "localhost:2345",
-      "miDebuggerPath": "${cmake.buildRoot}/arm-none-eabi-gdb.exe",
-      "timeout": 10000,
-      "setupCommands": [
-        {
-          "description": "Enable pretty-printing for gdb",
-          "text": "-enable-pretty-printing",
-          "ignoreFailures": true
-        },
-        {
-          "description": "mGBA",
-          "text": "shell start /b \"\" \"${env.MGBA}/mGBA.exe\" -g ${cmake.buildRoot}/rom.elf",
-          "ignoreFailures": false
-        }
-      ]
-    },
-    {
-      "type": "cppdbg",
-      "project": "CMakeLists.txt",
-      "projectTarget": "rom.elf",
-      "name": "Attach Debugger",
-      "MIMode": "gdb",
-      "program": "${cmake.buildRoot}/rom.elf",
-      "targetArchitecture": "arm",
-      "miDebuggerServerAddress": "localhost:2345",
-      "miDebuggerPath": "${cmake.buildRoot}/arm-none-eabi-gdb.exe",
-      "request": "attach",
-      "setupCommands": [
-        {
-          "description": "Enable pretty-printing for gdb",
-          "text": "-enable-pretty-printing",
-          "ignoreFailures": true
-        }
-      ]
-    }
-  ]
-}
