--- sample_project/tasks.vs.json
@@ -1,35 +1,0 @@
-{
-  "version": "0.2.1",
-  "tasks": [
-    {
-      "workingDirectory": "${env.MGBA}",
-      "taskLabel": "mGBA (ROM) GDB server",
-      "appliesTo": "*",
-      "type": "launch",
-      "command": "mGBA",
-      "args": [
-        "-g ${cmake.buildRoot}/rom.elf"
-      ]
-    },
-    {
-      "workingDirectory": "${env.MGBA}",
-      "taskLabel": "mGBA (ROM)",
-      "appliesTo": "*",
-      "type": "launch",
-      "command": "mGBA",
-      "args": [
-        "${cmake.buildRoot}/rom.gba"
-      ]
-    },
-    {
-      "workingDirectory": "${env.MGBA}",
-      "taskLabel": "mGBA (Multiboot)",
-      "appliesTo": "*",
-      "type": "launch",
-      "command": "mGBA",
-      "args": [
-        "${cmake.buildRoot}/multiboot.gba"
-      ]
-    }
-  ]
-}
