+++ samples/cplusplus/.gitignore
@@ -1,0 +1,6 @@
+cmake-build-debug/
+cmake-build-debug-clang/
+cmake-build-release/
+cmake-build-release-clang/
+build/
+!.idea/
