+++ samples/cplusplus/.idea/runConfigurations/mGBA_Debug_Clang.xml
@@ -1,0 +1,8 @@
+<component name="ProjectRunConfigurationManager">
+  <configuration default="false" name="mGBA Debug Clang" type="com.jetbrains.cidr.embedded.customgdbserver.type" factoryName="com.jetbrains.cidr.embedded.customgdbserver.factory" PROGRAM_PARAMS="-g cplusplus.elf" REDIRECT_INPUT="false" WORKING_DIR="file://$PROJECT_DIR$/cmake-build-debug-clang" PASS_PARENT_ENVS_2="true" PROJECT_NAME="cplusplus" TARGET_NAME="cplusplus" CONFIG_NAME="Debug" version="1" RUN_TARGET_PROJECT_NAME="cplusplus" RUN_TARGET_NAME="cplusplus">
+    <custom-gdb-server version="1" gdb-connect="localhost:2345" executable="E:/mGBA/mGBA.exe" warmup-ms="0" download-type="NONE" reset-cmd="monitor reset" reset-after-download="true" />
+    <method v="2">
+      <option name="CLION.COMPOUND.BUILD" enabled="true" />
+    </method>
+  </configuration>
+</component>
