+++ samples/cplusplus/.vscode/launch.json
@@ -1,0 +1,35 @@
+{
+  // Use IntelliSense to learn about possible attributes.
+  // Hover to view descriptions of existing attributes.
+  // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
+  "version": "0.2.0",
+  "configurations": [
+    {
+      "name": "(gdb) Launch",
+      "type": "cppdbg",
+      "request": "launch",
+      "program": "${workspaceFolderBasename}.elf",
+      "args": [],
+      "stopAtEntry": false,
+      "cwd": "${workspaceFolder}/build/",
+      "environment": [],
+      "externalConsole": false,
+      "MIMode": "gdb",
+      "miDebuggerServerAddress": "localhost:2345",
+      "windows": {
+        "miDebuggerPath": "${workspaceFolder}/../../arm-gnu-toolchain/bin/arm-none-eabi-gdb.exe",
+        "setupCommands": [
+            {
+            "description": "Enable pretty-printing for gdb",
+            "ignoreFailures": true,
+                "text": "file ${workspaceFolder}/build/${workspaceFolderBasename}.elf -enable-pretty-printing"
+            },
+            {
+              "description": "mGBA",
+              "text": "shell start /b \"\" \"${env:MGBA}/mGBA.exe\" -g ${workspaceFolder}/build/${workspaceFolderBasename}.elf",
+              "ignoreFailures": false
+            }]
+      }
+    }
+  ]
+}
