+++ samples/tonclib/.idea/runConfigurations/Debug.xml
@@ -1,0 +1,8 @@
+<component name="ProjectRunConfigurationManager">
+  <configuration default="false" name="Debug" type="com.jetbrains.cidr.embedded.customgdbserver.type" factoryName="com.jetbrains.cidr.embedded.customgdbserver.factory" PROGRAM_PARAMS="-g tonclib.elf" REDIRECT_INPUT="false" WORKING_DIR="file://$PROJECT_DIR$/cmake-build-debug" PASS_PARENT_ENVS_2="true" PROJECT_NAME="tonclib" TARGET_NAME="tonclib" CONFIG_NAME="Debug" version="1" RUN_TARGET_PROJECT_NAME="tonclib" RUN_TARGET_NAME="tonclib">
+    <custom-gdb-server version="1" gdb-connect="localhost:2345" executable="E:/mGBA/mGBA" warmup-ms="0" download-type="UPDATED_ONLY" reset-cmd="monitor reset" reset-after-download="true" />
+    <method v="2">
+      <option name="CLION.COMPOUND.BUILD" enabled="true" />
+    </method>
+  </configuration>
+</component>
