+++ samples/tonclib/main.c
@@ -1,0 +1,5 @@
+#include <tonc.h>
+
+int main( int argc, char * argv[] ) {
+    return 0;
+}
